Further Column Trait Tidy Up
